整理分析器文件夹结构
